Removed Facebook message hint, added email contact.
As of June 22, 2016, the Facebook messaging for the Parish has been closed due to diverging communications between the actual office operations and the expected details listed on the page. Various issues cannot be handled by the page itself (i.e., phone issues at the office) and nearly every message to the Parish has been forwarded either through email or the phone.